requirements.txt

Django>=5.0,<6.1
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
whitenoise
gunicorn
requests
bcrypt
